% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_markers.R
\name{annotate_markers}
\alias{annotate_markers}
\title{Annotate Markers}
\usage{
annotate_markers(
  top10_genes_data = NULL,
  species = NULL,
  tissue_class = NULL,
  data_source = NULL
)
}
\arguments{
\item{top10_genes_data}{A named list where names are cluster identifiers and values are vectors of top 10 gene symbols for each cluster.}

\item{species}{A character string specifying the species. It can be either "Human" or "Mouse".}

\item{tissue_class}{A character string specifying the tissue class.}

\item{data_source}{A character string specifying the data source to view. It can be either "CellMarker", "CellMarker2.0", or "PanglaoDB".}
}
\value{
A list with two components: "all_filtered_data" which is a list where each element is a data frame of filtered data for each cluster, and "all_filtered_data_df" which is a data frame combining all the filtered data.
}
\description{
This function filters a preloaded dataset based on provided top 10 genes for each cluster, species, and tissue class. It also allows to view the data from a particular data source.
}
\examples{
\dontrun{
  data <- list("cluster1" = c("gene1", "gene2"), "cluster2" = c("gene3", "gene4"))
  result <- annotate_markers(data, "Human", "Brain")
  view_data <- annotate_markers(data_source = "CellMarker")
}
}
