% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DXMarkers_dotplots.R
\name{DXMarkers_dotplots}
\alias{DXMarkers_dotplots}
\title{Generate DotPlots from a Seurat object and save as PDF}
\arguments{
\item{pbmc_data}{a Seurat object}

\item{csv_file}{a string, path to the csv file containing cell type and gene information}

\item{directory}{a string, path to the directory where the PDFs will be saved}
}
\description{
This function generates DotPlots for each cell type from a Seurat object, using the genes specified in a csv file, and saves the plots as PDFs in a specified directory.

This function generates DotPlots for each cell type from a Seurat object, using the genes specified in a csv file, and saves the plots as PDFs in a specified directory.
}
\examples{
DXMarkers_dotplots(pbmc3k.final, "confirm_SCINA_markers.csv", "C:\\\\Users\\\\Wandering\\\\Desktop\\\\测试")
DXMarkers_dotplots(pbmc3k.final, "confirm_SCINA_markers.csv", "C:\\\\Users\\\\Wandering\\\\Desktop\\\\测试")
}
