% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DXMarkers_heatmap.R
\name{DXMarkers_heatmap}
\alias{DXMarkers_heatmap}
\title{Generate Heatmap for DX Markers with log(*+1) transformed expression}
\usage{
DXMarkers_heatmap(
  scRNA,
  DXMarkers_result,
  output_file,
  width = 800,
  height = 800,
  res = 120
)
}
\arguments{
\item{scRNA}{An object containing single-cell RNA sequencing data. It can be Seurat, SingleCellExperiment, or other supported formats.}

\item{DXMarkers_result}{A list containing gene signatures for different cell types.}

\item{output_file}{Path to the output PNG file where the heatmap will be saved.}

\item{width}{Width of the output PNG image in pixels. Default is 800 pixels.}

\item{height}{Height of the output PNG image in pixels. Default is 800 pixels.}

\item{res}{Resolution (dots per inch) for the output PNG image. Default is 120 dpi.}
}
\description{
This function generates a heatmap for DX Markers from single-cell RNA sequencing data (scRNA) with log(*+1) transformed expression values, and saves it as a PNG file.
}
\examples{

# 载入所需的包
library(Seurat)

# 加载单细胞RNA数据对象
# Replace "scRNA" with the actual object containing your scRNA data
scRNA <- your_scRNA_data

# 定义gene signatures的列表
# Replace "DXMarkers_result" with the actual list of gene signatures
DXMarkers_result <- your_DXMarkers_result

# 指定输出的PNG文件名
output_file <- "heatmap_output.png"

# 生成热图
DXMarkers_heatmap(scRNA, DXMarkers_result, output_file)
}
