% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCINA_markers.R
\name{SCINA_markers}
\alias{SCINA_markers}
\title{Extract Marker Genes for Each Cell Type}
\arguments{
\item{annotate_markers_result}{The result of the 'annotate_markers' function.}
}
\value{
A data frame where each column represents a cell type and each row in that column represents a gene. The name of the column is the cell type and the entries in that column are the genes.
}
\description{
This function extracts the marker genes for each cell type from the result of the 'annotate_markers' function.
}
\examples{
\dontrun{
  data <- list("cluster1" = c("gene1", "gene2"), "cluster2" = c("gene3", "gene4"))
  annotate_result <- annotate_markers(data, "Human", "Brain")
  markers <- SCINA_markers(annotate_result)
}
}
