% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_to_matrix.R
\name{Seurat_to_matrix}
\alias{Seurat_to_matrix}
\title{Seurat_to_matrix}
\usage{
Seurat_to_matrix(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object.}
}
\value{
A gene expression matrix where each row represents a gene and each column represents a cell.
}
\description{
The function is designed to extract the gene expression matrix from a Seurat object.
It will create a new matrix where each row represents a gene, and each column represents a cell.
Note: The input should be a Seurat object.
}
\examples{
\dontrun{
seurat_data <- Read10X(data.dir = "../data")
seurat_obj <- CreateSeuratObject(counts = seurat_data, project = "seurat_data", min.cells = 3, min.features = 200)
expression_matrix <- Seurat_to_matrix(seurat_obj)
}
}
