% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_data_source.R
\name{view_data_source}
\alias{view_data_source}
\title{View Data Source}
\usage{
view_data_source(data_source)
}
\arguments{
\item{data_source}{A character string specifying the data source to view. It can be either "CellMarker", "CellMarker2.0", or "PanglaoDB".}
}
\value{
A dataframe containing the dataset from the selected data source.
}
\description{
This function returns the dataset from a particular data source.
}
\examples{
\dontrun{
  view_data <- view_data_source(data_source = "CellMarker")
}
}
