% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_markers.R
\name{annotate_markers}
\alias{annotate_markers}
\title{Annotate Cell and Gene Markers from Multiple Databases}
\usage{
annotate_markers(top10_genes_data, species, tissue_class)
}
\arguments{
\item{species}{species name, 'Mouse' or 'Human'}

\item{tissue_class}{tissue class to filter}

\item{top10genes_path}{file path of the top 10 genes csv file}
}
\value{
a dataframe of the filtered data
}
\description{
This function annotates cell and gene markers from multiple databases.
}
