% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_genes_wide.R
\name{reshape_genes_wide}
\alias{reshape_genes_wide}
\title{重塑基因数据为宽格式并排序，并保存为CSV文件}
\usage{
reshape_genes_wide(sorted_genes)
}
\arguments{
\item{sorted_genes}{经过排序后的基因数据框}
}
\value{
一个包含重塑后的基因数据框的list
}
\description{
该函数将经过排序后的基因数据重塑成宽格式，每个cluster为一列，并保存结果至CSV文件中。
}
